Added majority of events
